requirements.txt
ipykernel
pandas-2.3.3
geopandas-1.1.1
osmnx-2.0.6
folium-0.20.0